feat(servicenow): use original servicenow fields to be alignged with query filter
