+++ workspaces/servicenow/plugins/servicenow/src/types.ts
@@ -14,6 +14,15 @@
  * limitations under the License.
  */
 
+export type IncidentsData = {
+  number: string;
+  shortDescription: string;
+  description: string;
+  sysCreatedOn: string;
+  priority: number;
+  incidentState: number;
+};
+
 export enum IncidentTableFieldEnum {
   Number = 'number',
   ShortDescription = 'short_description',
